Revert "Python3 anuga py3"

+++ anuga/file/mux.py
@@ -1,9 +1,7 @@
 """
     Read a mux2 file.
 """
-from __future__ import absolute_import
 
-from builtins import range
 from anuga.utilities.numerical_tools import ensure_numeric
 import numpy as num     
         
@@ -37,7 +35,7 @@ def read_mux2_py(filenames,
                         extracted
     """
 
-    from .urs_ext import read_mux2
+    from urs_ext import read_mux2
 
     numSrc = len(filenames)
 
@@ -53,7 +51,7 @@ def read_mux2_py(filenames,
         weights = num.ones(numSrc)
 
     if permutation is None:
-        permutation = ensure_numeric([], num.int)
+        permutation = ensure_numeric([], num.float)
 
     # Call underlying C implementation urs2sts_ext.c
     data = read_mux2(numSrc, filenames, weights, file_params,
